requirements.txt
# Core LLM & RAG logic
langchain>=0.1.15
openai>=1.30.1
tiktoken>=0.6.0

# Vector search
faiss-cpu>=1.7.4
sentence-transformers>=2.6.1  # Optional: if you use BGE/MiniLM embeddings

# PDF processing
PyMuPDF>=1.23.22

# Web backend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Frontend UI
streamlit>=1.35.0

# Environment & config
python-dotenv>=1.0.1

# Typing & utilities
pydantic>=2.7.1
httpx>=0.27.0

# Optional testing
pytest>=8.2.1